+++ src/App.js
@@ -4,7 +4,7 @@ import AuthPage from './components/Auth/AuthPage';
 import Home from './components/Home/Home';
 import Profile from './components/Profile/Profile';
 import Search from './components/Search/Search';
-import PrivateRoute from './components/Auth/PrivateRoute';
+import PrivateRoute from './components/Auth/PrivateRoute/PrivateRoute';
 import 'react-toastify/dist/ReactToastify.css';
 import { ToastContainer } from 'react-toastify';
